Changed name to python server and began adding functions.
